Add async base we can derive from, to be used for members and media too
